data_injector/requirements.txt
requests==2.28.2
pandas==2.0.1
mysql-connector-python
os-win==5.9.0
kaggle==1.5.13